+++ flow/defs.js
@@ -1,0 +1,6 @@
+declare module "react-test-renderer/shallow" {
+  declare var exports: any;
+}
+declare module "react-element-to-jsx-string" {
+  declare var exports: any;
+}