+++ src/__tests__/__fixtures__/fn.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function FakeComponent() {}
+
+export default function FunctionalComponentToTest(props) {
+  return <FakeComponent fixedProp={"some-value"} {...props} />;
+}