+++ src/index.js
@@ -1,23 +1,31 @@
-import { createRenderer } from "react-addons-test-utils"
-import reactElementToJSXString from "react-element-to-jsx-string"
+import ReactShallowRenderer from "react-test-renderer/shallow";
+import reactElementToJSXString from "react-element-to-jsx-string";
 
 export function noop() {}
 
-export const JSX = reactElementToJSXString
+export const JSX = reactElementToJSXString;
 
-export const render = (jsx, cb = noop, context) => {
+type cbType = (render: Object, renderer: Object) => void;
+
+export const render = (jsx: Object, cb?: cbType | Object, context?: Object) => {
   if (typeof cb === "object") {
-    context = cb
-    cb = noop
+    context = cb;
+    cb = undefined;
   }
-  const renderer = createRenderer()
-  const render = renderer.render(jsx, context)
+  const renderer = new ReactShallowRenderer();
+  const render = renderer.render(jsx, context);
 
   // here you can execute whatever you want on the renderer
   // eg: (render) => render.props.onClick({ preventDefault: noop })
-  cb(render, renderer)
+  if (typeof cb === "function") {
+    cb(render, renderer);
+  }
 
-  return renderer.getRenderOutput()
-}
+  return renderer.getRenderOutput();
+};
 
-export const renderJSX = (...args) => JSX(render(...args))
+export const renderJSX = (
+  jsx: Object,
+  cb?: cbType | Object,
+  context?: Object
+) => JSX(render(jsx, cb, context));