+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0 - 2017-06-08
+
+- Upgrade `react-element-to-jsx-string` to ^10.0.0
+- Replace `react-addons-test-utils` by `react-test-renderer`
+- Replace AVA to Jest in examples
+
 # 1.1.0 - 2016-04-25
 
 - Added: support for context as an additional of render/renderJSX